Fully functional 'enable local services' link in boot dashboard

+++ plugins/org.springframework.ide.eclipse.boot.dash/src/org/springframework/ide/eclipse/boot/dash/model/local/AutoCloudCliInstaller.java
@@ -8,12 +8,17 @@
  * Contributors:
  *     Spring IDE Developers - initial API and implementation
  *******************************************************************************/
-package org.springframework.ide.eclipse.boot.core.cli.install;
+package org.springframework.ide.eclipse.boot.dash.model.local;
 
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.core.runtime.IStatus;
 import org.eclipse.core.runtime.Status;
 import org.eclipse.core.runtime.jobs.Job;
+import org.springframework.ide.eclipse.boot.core.cli.install.AutoInstallDescription;
+import org.springframework.ide.eclipse.boot.core.cli.install.CloudCliInstall;
+import org.springframework.ide.eclipse.boot.core.cli.install.IBootInstall;
+import org.springframework.ide.eclipse.boot.core.cli.install.IBootInstallExtension;
+import org.springframework.ide.eclipse.boot.dash.model.UserInteractions;
 import org.springframework.ide.eclipse.boot.util.Log;
 import org.springsource.ide.eclipse.commons.livexp.util.ExceptionUtil;
 
@@ -30,21 +35,26 @@ public class AutoCloudCliInstaller {
 		this.bootInstall = bootInstall;
 	}
 
-	public void performInstall() {
-		InstallBootCliExtensionJob installCloudCliJob = new InstallBootCliExtensionJob("Auto install Spring Cloud CLI", CloudCliInstall.class);
-		installCloudCliJob.schedule();
-		long waitTime = 200L;
-		for (long waited = 0L; installCloudCliJob.getState() != Job.NONE && waited < 30000L; waited+=waitTime) {
-			try {
-				Thread.sleep(waitTime);
-			} catch (InterruptedException e) {
-				// ignore
+	public void performInstall(UserInteractions ui) {
+		AutoInstallDescription autoInstallation = bootInstall.checkAutoInstallable(CloudCliInstall.class);
+		if (!autoInstallation.isPossible) {
+			ui.errorPopup("Auto installation of Spring Cloud CLI not possible", autoInstallation.message);
+		} else if (ui.confirmOperation("Confirm Installation of Spring Cloud CLI?", autoInstallation.message)) {
+			InstallBootCliExtensionJob installCloudCliJob = new InstallBootCliExtensionJob("Auto install Spring Cloud CLI", CloudCliInstall.class);
+			installCloudCliJob.schedule();
+			long waitTime = 200L;
+			for (long waited = 0L; installCloudCliJob.getState() != Job.NONE && waited < 30000L; waited+=waitTime) {
+				try {
+					Thread.sleep(waitTime);
+				} catch (InterruptedException e) {
+					// ignore
+				}
 			}
+			if (installCloudCliJob.getState() != Job.NONE) {
+				Log.error("Timed out waiting for Spring Cloud CLI to be installed");
+			}
+			bootInstall.refreshExtensions();
 		}
-		if (installCloudCliJob.getState() != Job.NONE) {
-			Log.error("Timed out waiting for Spring Cloud CLI to be installed");
-		}
-		bootInstall.refreshExtensions();
 	}
 
 	private class InstallBootCliExtensionJob extends Job {
